+++ cotests/__init__.py
@@ -1,4 +1,4 @@
 from .batch import bench_batch, test_batch
 from .obj import CoTestCase
-from .bench import CoTestGroup, test_groups
+from .group import CoTestGroup, test_groups
 from .module import test_module
