--- cotests/bench/__init__.py
@@ -1,2 +1,0 @@
-from .case import AbstractCoCase
-from .group import CoTestGroup, test_groups
