+++ cotests/case/abstract.py
