+++ cotests/cases/abstract.py
