+++ cotests/group/ctx.py
@@ -2,9 +2,9 @@ from time import perf_counter
 from contextlib import contextmanager
 from typing import List, Iterable
 
-from ..cases import AbstractTestCase
+from cotests.cases import AbstractTestCase
 from cotests.exceptions import CoException
-from ..utils import get_level_prefix, format_sec_metrix, print_test_results
+from cotests.utils import get_level_prefix, format_sec_metrix, print_test_results
 
 
 class AbstractTestGroup(AbstractTestCase):
